# example/cmd/server/routers/public.go
package routers

import (
	dtoAuth "example/internal/auth-service/dto/request"
	dtoRefresh "example/internal/refresh-service/dto/request"

	w_handl "github.com/DVV-15324/witches/pkg/core/handler"
	w_resp "github.com/DVV-15324/witches/pkg/core/response"
	"github.com/ulule/limiter/v3"
)

// En: This function groups routes that don't need to authenticate with each other.
//
// Vi: Hàm nhóm các routes không cần xác thực với nhau.
func setupPublicRoutes(gen *w_handl.SwaggerGenerator, services *HandleServices, limiterRate *limiter.Rate) {
	// En: Login
	// Vi: Đăng nhập
	gen.POST("/v1/auth/login").
		Summary("User login").
		Description("Authenticate user and return JWT token").
		Tags("auth").
		HeaderParam("User-Agent", "User-Agent", false).
		HeaderParam("Accept-Language", "Preferred language (e.g., vi-VN, en-US)", false).
		HeaderParam("X-Timezone", "Timezone (e.g., Asia/Ho_Chi_Minh)", false).
		RateLimit(*limiterRate).
		Body(dtoAuth.Login{}, "Login credentials").
		Response(200, w_resp.AppResponse{}, "Login successful").
		Response(401, w_resp.AppResponse{}, "Unauthorized").
		Handler(services.HandleAuth.HandleLogin()).
		Build()

	// En: Login with Google
	// Vi: Đăng nhập với Google
	gen.POST("/v1/auth/login-google").
		Summary("Google login").
		Description("Authenticate user using Google OAuth2 ID token").
		Tags("auth").
		HeaderParam("User-Agent", "User-Agent", false).
		HeaderParam("Accept-Language", "Preferred language (e.g., vi-VN, en-US)", false).
		HeaderParam("X-Timezone", "Timezone (e.g., Asia/Ho_Chi_Minh)", false).
		RateLimit(*limiterRate).
		Body(dtoAuth.GoogleLoginRequest{}, "Google ID token").
		Response(200, w_resp.AppResponse{}, "Login successful").
		Response(401, w_resp.AppResponse{}, "Unauthorized").
		Handler(services.HandleAuth.HandleGoogleLogin()).
		Build()

	// En: Register
	// Vi: Đăng kí
	gen.POST("/v1/auth/register").
		Summary("User registration").
		Description("Register a new user account").
		Tags("auth").
		HeaderParam("User-Agent", "User-Agent", false).
		HeaderParam("Accept-Language", "Preferred language (e.g., vi-VN, en-US)", false).
		HeaderParam("X-Timezone", "Timezone (e.g., Asia/Ho_Chi_Minh)", false).
		RateLimit(*limiterRate).
		Body(dtoAuth.Register{}, "User registration data").
		Response(201, w_resp.AppResponse{}, "Registration successful").
		Response(400, w_resp.AppResponse{}, "Bad request").
		Handler(services.HandleAuth.HandleRegister()).
		Build()

	// En: Refresh token
	// Vi: Refresh token
	gen.POST("/v1/auth/refresh").
		Summary("Refresh access token").
		Description("Get new access token using refresh token").
		Tags("auth").
		HeaderParam("User-Agent", "User-Agent", false).
		HeaderParam("Accept-Language", "Preferred language (e.g., vi-VN, en-US)", false).
		HeaderParam("X-Timezone", "Timezone (e.g., Asia/Ho_Chi_Minh)", false).
		RateLimit(*limiterRate).
		Body(dtoRefresh.RefreshTokenRequest{}, "Refresh token").
		Response(200, w_resp.AppResponse{}, "Token refreshed successfully").
		Response(401, w_resp.AppResponse{}, "Invalid or expired refresh token").
		Handler(services.HandleRefresh.HandleRefreshToken()).
		Build()
}